Moved static functions from SourceMapCache to FitUtils
