+++ Likelihood/SourceMapCache.h
@@ -45,129 +45,6 @@ namespace Likelihood {
 
      /* ------------- Static Utility Functions -------------------- */
 
-     /* Get the model counts contribution to a particular pixel 
-
-	srcMap     : The SourceMap for the source in question
-	spec       : The specturm (or spectral derivative) for the source in question
-	jmin       : The unrolled index for the source map at the lower edge of the energy bin
-	jmax       : The unrolled index for the source map at the upper edge of the energy bin
-	kref       : The energy bin
-	emin       : The energy at the low edge of the bin
-	emax       : The energy at the high edge of the bin
-	log_ratio  : The log of the energy ratios in the bin
-
-	returns the contribution
-     */
-     static double model_counts_contribution(SourceMap& srcMap,
-					     const std::vector<double> & spec,
-					     size_t jmin,
-					     size_t jmax,
-					     size_t kref,
-					     const double& emin,
-					     const double& emax,
-					     const double& log_ratio);
-
-    /* Get the model counts contribution to a energy layer
-
-	npred_vals     : The total npred for that energy layer
-	npred_weights  : The weight factors for that energy layer
-	spec       : The specturm (or spectral derivative) for the source in question
-	kref       : The energy bin
-	emin       : The energy at the low edge of the bin
-	emax       : The energy at the high edge of the bin
-	log_ratio  : The log of the energy ratios in the bin
-	counts     : Filled with the counts contribution
-	counts_wt  : Filled with the weighted counts contribution
-     */
-     static void npred_contribution(const std::vector<double>& npred_vals,
-				    const std::vector<std::pair<double,double> >& npred_weights,
-				    const std::vector<double>& spec,
-				    size_t kref,
-				    const double& emin,
-				    const double& emax,
-				    const double& log_ratio,
-				    double& counts,
-				    double& counts_wt);
-
-
-     /* Add (or subtract) the counts for a source onto a vector 	
-	This is used by several functions.
-
-	modelCounts: The vector being added to.
-	srcMap     : The SourceMap for the source in question
-	npix       : Number of pixel in the map, used for indexing
-	filledPixels : Vector with the indices of the filled pixels,
-	dataCache  : Object with info about the binning
-	use_edisp_val : Apply the energy dispersion
-	subtract   : If true, subtract from the vector.  	
-     */     
-
-     static void addSourceCounts_static(std::vector<double> & modelCounts,
-					SourceMap& srcMap,
-					const BinnedCountsCache& dataCache,
-					bool use_edisp_val,
-					bool subtract);
- 
-     /* Add (or subtract) the counts for a fxied source onto the vectors that 
-	collect that info for fixed sources.
-	This is used by several functions.
-
-	fixed_counts_spec : The vector being added to.
-	fixed_counts_spec_wt : The vector being added to.
-	fixed_counts_spec_edisp : The vector being added to.
-	fixed_counts_spec_edisp_wt : The vector being added to.
-	srcMap     : The SourceMap for the source in question
-	dataCache  : Object with info about the binning
-	use_edisp_val : Apply the energy dispersion
-	subtract   : If true, subtract from the vector.  	
-     */     
-     static void addFixedNpreds_static(std::vector<double>& fixed_counts_spec,
-				       std::vector<double>& fixed_counts_spec_wt,
-				       std::vector<double>& fixed_counts_spec_edisp,
-				       std::vector<double>& fixed_counts_spec_edisp_wt,
-				       SourceMap& srcMap,
-				       const BinnedCountsCache& dataCache,
-				       bool use_edisp_val,
-				       bool subtract);
-
-#ifndef SWIG
-     /* Add (or subtract) the contributions to the derivatives from a single source.
-
-	posDerivs : The vector being added to.
-	negDerivs : The vector being added to.
-	freeIndex : The overall index of the first parameter for this source
-	srcMap     : The SourceMap for the source in question
-	dataCache  : Object with info about the binning
-	use_edisp_val : Apply the energy dispersion
-     */     
-     static void addFreeDerivs_static(std::vector<Kahan_Accumulator>& posDerivs,
-				      std::vector<Kahan_Accumulator>& negDerivs,
-				      long freeIndex,
-				      SourceMap& srcMap,
-				      const std::vector<float> & data,
-				      const std::vector<double>& model, 
-				      const BinnedCountsCache& dataCache,
-				      bool use_edisp_val,
-				      size_t kmin, size_t kmax);
-#endif //SWIG
-
-          
-    /* Add (or subtract) the counts for a source onto a vector.
-       This version loops over all the pixels, not just the filled ones.
-
-	modelMap   : The vector being added to.
-	srcMap     : The SourceMap for the source in question
-	dataCache  : Object with info about the binning
-	use_mask   : Use the Weights mask
-	use_edisp_val : Apply the energy dispersion
-     */     
-     static void updateModelMap_static(std::vector<float> & modelMap,
-				       SourceMap& srcMap,
-				       const BinnedCountsCache& dataCache,				       
-				       bool use_mask,
-				       bool use_edisp_val);
-
-     
 
    public:
