+++ src/FitUtils.cxx
@@ -17,6 +17,7 @@
 #include "gsl/gsl_vector.h"
 #include "gsl/gsl_linalg.h"
 
+#include "Likelihood/Drm.h"
 #include "Likelihood/Source.h"
 #include "Likelihood/SourceMap.h"
 #include "Likelihood/SourceModel.h"
@@ -1718,9 +1719,6 @@ namespace Likelihood {
     }
 
 
-    
-
-
     double logLikePoisson(std::vector<float>::const_iterator data_start, 
 			  std::vector<float>::const_iterator data_stop,
 			  std::vector<float>::const_iterator model_start,
@@ -1848,6 +1846,291 @@ namespace Likelihood {
       return retVal;
     }
 
+    double model_counts_contribution(SourceMap& srcMap,
+				     const std::vector<double> & spec,
+				     size_t jmin,
+				     size_t jmax,
+				     size_t kref,
+				     const double& emin,
+				     const double& emax,
+				     const double& log_ratio) {    
+      double y1 = srcMap[jmin]*spec[kref];
+      double y2 = srcMap[jmax]*spec[kref+1];
+      return FitUtils::pixelCounts_loglogQuad(emin, emax, y1, y2, log_ratio);
+    }
+
+    void npred_contribution(const std::vector<double>& npred_vals,
+			    const std::vector<std::pair<double,double> >& npred_weights,
+			    const std::vector<double>& spec,
+			    size_t kref,
+			    const double& emin,
+			    const double& emax,
+			    const double& log_ratio,
+			    double& counts,
+			    double& counts_wt) {
+      double npred_0 = npred_vals[kref] * spec[kref];
+      double npred_1 = npred_vals[kref+1] * spec[kref+1];
+      double npred_wt_0 = npred_0 * npred_weights[kref].first;
+      double npred_wt_1 = npred_1 * npred_weights[kref].second;
+      counts = FitUtils::pixelCounts_loglogQuad(emin, emax, npred_0, npred_1, log_ratio);
+      counts_wt = FitUtils::pixelCounts_loglogQuad(emin, emax, npred_wt_0, npred_wt_1, log_ratio);
+    }
+
+
+    double model_counts_edisp(SourceMap& srcMap,
+			      const std::vector<double> & spec,
+			      const BinnedCountsCache& dataCache,
+			      size_t ipix,
+			      size_t kref,
+			      size_t kmin,
+			      size_t kmax) {
+      if ( kmin >= kmax ) {
+	return 0.;
+      }
+      
+      const Drm* drm = srcMap.cached_drm();
+      if ( drm == 0 ) {
+	return 0.;
+      }
+      
+      std::vector<double> srcMapValues(kmax-kmin);
+      srcMap.extractPixelValues(srcMapValues, ipix, kmin, kmax);
+      const std::vector<double> & evals = dataCache.energies();
+      const std::vector<double> & log_ratios = dataCache.log_energy_ratios();
+      const std::vector<double> & edisp_col = drm->col(kref);
+      
+      size_t idx(0);
+      size_t khi(kmin+1);
+      double ret_val(0.);
+      for ( size_t k(kmin); k < kmax; k++, khi++, idx++ ) {
+	double y1 = srcMapValues[k] * spec[k];
+	double y2 = srcMapValues[khi] * spec[khi];
+	ret_val += edisp_col[k] * pixelCounts_loglogQuad(evals[k], evals[khi], y1, y2, log_ratios[k]);
+      }
+
+      return ret_val;
+    }
+
+
+    void addSourceCounts(std::vector<double> & modelCounts,
+			 SourceMap& srcMap,
+			 const BinnedCountsCache& dataCache,
+			 bool use_edisp_val,
+			 bool subtract) {
+      
+      const Drm_Cache* drm_cache = use_edisp_val ? srcMap.drm_cache() : 0;
+      double my_sign = subtract ? -1.0 : 1.0;
+      int kref(-1);
+      const std::vector<double> & spec = srcMap.specVals();
+      size_t nebins = dataCache.num_ebins();
+      size_t npix = dataCache.num_pixels();
+      const std::vector<size_t>& pix_ranges = dataCache.firstPixels();
+      for (size_t k(0); k < nebins; k++ ) {
+	size_t j_start = pix_ranges[k];
+	size_t j_stop = pix_ranges[k+1];      
+	double emin(dataCache.energies()[k]);
+	double emax(dataCache.energies()[k+1]);
+	double log_ratio(dataCache.log_energy_ratios()[k]);
+	int kref(-1);
+	double xi(1.);
+	if (use_edisp_val) {
+	  xi = drm_cache->get_correction(k,kref);
+	  if ( kref < 0 ) { 
+	    // Still have true counts in this energy bin, just use k as kref
+	    kref = k;
+	  } 
+	} else { // No energy disperion
+	  kref = k;
+	}      
+	for (size_t j(j_start); j < j_stop; j++) {
+	  size_t jmin(dataCache.filledPixels()[j]);
+	  size_t jmax(jmin + npix);
+	  if ( kref != k ) {
+	    size_t ipix(jmin % npix);
+	    jmin = kref*npix + ipix;
+	    jmax = jmin + npix;
+	  }
+	  double counts = model_counts_contribution(srcMap, spec, jmin, jmax, kref, emin, emax, log_ratio);      
+	  double addend = my_sign*xi*counts;
+	  modelCounts[j] += addend;
+	}
+      }
+    }
+
+    void addFixedNpreds(std::vector<double>& fixed_counts_spec,
+			std::vector<double>& fixed_counts_spec_wt,
+			std::vector<double>& fixed_counts_spec_edisp,
+			std::vector<double>& fixed_counts_spec_edisp_wt,
+			SourceMap& srcMap,
+			const BinnedCountsCache& dataCache,
+			bool use_edisp_val,
+			bool subtract) {
+      
+      const Drm_Cache* drm_cache = use_edisp_val ? srcMap.drm_cache() : 0;
+      bool has_wts = srcMap.weights() != 0;
+      double factor = subtract ? -1.0 : 1.0;
+      const std::vector<double>& npred_vals = srcMap.npreds();
+      const std::vector<std::pair<double,double> >& npred_weights = srcMap.npred_weights();
+      const std::vector<double>& engs = dataCache.energies();
+      const std::vector<double>& log_energy_rats = dataCache.log_energy_ratios();
+      const std::vector<double>& spec = srcMap.specVals(); 
+      
+      size_t ne(dataCache.num_ebins());
+      
+      double counts(0.);
+      double counts_wt(0.);
+      
+      for (size_t k(0); k < ne; k++) {
+	double xi(1);
+	int kref(-1);
+	if (use_edisp_val) {
+	  xi = drm_cache->get_correction(k,kref,has_wts);
+	  if ( kref < 0 ) { 
+	    // Still have true counts in this energy bin, just use k as kref
+	    kref = k;
+	  } 
+	} else {
+	  kref = k;
+	}      
+	
+	npred_contribution(npred_vals, npred_weights, spec, kref, engs[k], engs[k+1], log_energy_rats[k], counts, counts_wt);
+	counts *= factor;
+	counts_wt *= factor;
+	
+	fixed_counts_spec[k] += counts;
+	fixed_counts_spec_wt[k] += counts_wt;
+	fixed_counts_spec_edisp[k] += xi*counts;
+	fixed_counts_spec_edisp_wt[k] += xi*counts_wt;
+      }
+    }
+
+    void addFreeDerivs(std::vector<Kahan_Accumulator>& posDerivs,
+		       std::vector<Kahan_Accumulator>& negDerivs,
+		       long freeIndex,
+		       SourceMap& srcMap,
+		       const std::vector<float> & data,
+		       const std::vector<double>& model, 
+		       const BinnedCountsCache& dataCache,
+		       bool use_edisp_val,
+		       size_t kmin, size_t kmax) {
+      
+      const Drm_Cache* drm_cache = use_edisp_val ? srcMap.drm_cache() : 0;
+      bool has_wts = srcMap.weights() != 0;
+      const std::vector< std::vector<double> > & specDerivs = srcMap.cached_specDerivs();
+      
+      // We need this stuff for the second term...
+      const std::vector<double> & npreds = srcMap.npreds();
+      const std::vector<std::pair<double,double> > & npred_weights =  srcMap.npred_weights();
+      
+      size_t npix = dataCache.num_pixels();
+      const std::vector<size_t>& pix_ranges = dataCache.firstPixels();    
+      
+      for (size_t k(kmin); k < kmax; k++ ) {
+	double emin(dataCache.energies()[k]);
+	double emax(dataCache.energies()[k+1]);
+	double log_ratio(dataCache.log_energy_ratios()[k]);
+	int kref(-1);
+	double xi(1.);
+	if (use_edisp_val) {
+	  xi = drm_cache->get_correction(k,kref,has_wts);
+	  if ( kref < 0 ) {
+	    // Still have true counts in this energy bin, just use kk as kref
+	    kref = k;
+	  }
+	} else {
+	  kref = k;
+	}
+	
+	// First term, derivate of n_obs log n_model = ( n_obs / n_model ) * ( d model / d param ) 
+	// loop over all the filled pixels
+	size_t j_start = pix_ranges[k];
+	size_t j_stop = pix_ranges[k+1];      
+	for (size_t j(j_start); j < j_stop; j++) {
+	  if ( model[j] <= 0. ) {
+	    continue;
+	  }
+	  size_t jmin(dataCache.filledPixels()[j]);
+	  size_t jmax(jmin + npix);
+	  if ( kref != k ) {
+	    size_t ipix(jmin % npix);
+	    jmin = kref*npix + ipix;
+	    jmax = jmin + npix;
+	  }
+	  long iparam(freeIndex);
+	  for (size_t i(0); i < specDerivs.size(); i++, iparam++) {
+	    double counts_deriv = model_counts_contribution(srcMap, specDerivs[i], jmin, jmax, kref, emin, emax, log_ratio);
+	    double my_deriv = xi * counts_deriv;
+	    double addend = (data[jmin]/model[j])*my_deriv;
+	    if (addend > 0) {
+	      posDerivs[iparam].add(addend);
+	    } else {
+	      negDerivs[iparam].add(addend);
+	    }
+	  } 	
+	}
+
+	// Second term, the derivatives of the nPreds. 
+	// Loop over the energy layers
+	long iparam2(freeIndex);
+	double counts_deriv(0.);
+	double counts_deriv_wt(0.);
+	for (size_t i2(0); i2 < specDerivs.size(); i2++, iparam2++) {	
+	  npred_contribution(npreds, npred_weights, specDerivs[i2], kref, emin, emax, log_ratio, counts_deriv, counts_deriv_wt);
+	  double addend = xi * counts_deriv_wt;
+	  if (-addend > 0) {
+	    posDerivs[iparam2].add(-addend);
+	  } else {
+	    negDerivs[iparam2].add(-addend);
+	  }
+	}
+      } // Loop on energy bins
+    }
+
+
+    void updateModelMap(std::vector<float> & modelMap,
+			SourceMap& srcMap,
+			const BinnedCountsCache& dataCache,				       
+			bool use_mask,
+			bool use_edisp_val) {
+      
+      const Drm_Cache* drm_cache = use_edisp_val ? srcMap.drm_cache() : 0;
+      const std::vector<double> & specVals = srcMap.specVals();  
+      const WeightMap* mask = use_mask ? srcMap.weights() : 0;
+      size_t npix = dataCache.num_pixels();
+      
+      int kref(-1);
+      for (size_t k(0); k < dataCache.num_ebins(); k++) {
+	double emin(dataCache.energies()[k]);
+	double emax(dataCache.energies()[k+1]);
+	double log_e_ratio(dataCache.log_energy_ratios()[k]);
+	for (size_t j(0); j < npix; j++) {
+	  size_t jmin(k*npix + j);
+	  size_t jmax(jmin + npix);
+	  // EAC, skip masked pixels
+	  if ( mask && 
+	       ( mask->model()[jmin] <= 0. ) ) continue;
+	  double xi(1.);
+	  int kref(-1);
+	  if (use_edisp_val) {
+	    xi = drm_cache->get_correction(k,kref);
+	    if ( kref < 0 ) { 
+	      // Still have true counts in this energy bin, just use k as kref
+	      kref = k;
+	    } else { // Need to use reference bin
+	      size_t ipix(jmin % npix);	
+	      jmin = kref*npix + ipix;
+	      jmax = jmin + npix;
+	    }
+	  } else { // No energy disperion
+	    xi = 1.0;
+	    kref = k;
+	  }      
+	  double counts = model_counts_contribution(srcMap, specVals, jmin, jmax, kref, emin, emax, log_e_ratio);  
+	  float val = xi*counts;
+	  modelMap[jmin] += val;
+	}
+      } 
+    }  
 
     void printVector(const std::string& name,
 		     const CLHEP::HepVector& vect) {
